+++ front-end/src/components/Header/Header.styled.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { up } from "styled-breakpoints";
-import { Text } from "../../components";
+import { Text } from "..";
 
 export const StyledHeader = styled.header`
   width: 100%;
