+++ front-end/src/components/Layout/Layout.stories.tsx
@@ -1,7 +1,8 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import styled from "styled-components";
 import { up } from "styled-breakpoints";
-import { Header, SubHeader } from "../../domains/Common";
+import Header from "../Header/Header";
+import SubHeader from "../SubHeader/SubHeader";
 import { ManagerNavigation } from "../../domains/Manager";
 import { GpuServerRegisterForm } from "../../features/gpuServer";
 import Layout from "./Layout";
