+++ front-end/src/components/index.ts
@@ -5,6 +5,7 @@ export { default as Dialog } from "./Dialog/Dialog";
 export { default as Dimmer } from "./Dimmer/Dimmer";
 export { default as Flicker } from "./Flicker/Flicker";
 export { default as GpuIcon } from "./GpuIcon/GpuIcon";
+export { default as Header } from "./Header/Header";
 export { default as Input } from "./Input/Input";
 export { default as Layout } from "./Layout/Layout";
 export { default as Loading } from "./Loading/Loading";
@@ -13,6 +14,7 @@ export { default as Portal } from "./Portal/Portal";
 export { default as Radio } from "./Radio/Radio";
 export { default as RadioGroup } from "./RadioGroup/RadioGroup";
 export { default as ServerIcon } from "./ServerIcon/ServerIcon";
+export { default as SubHeader } from "./SubHeader/SubHeader";
 export { default as Table } from "./Table/Table";
 export { default as Toast } from "./Toast/Toast";
 export { default as ToastProvider, useToast } from "./Toast/ToastProvider";
