--- front-end/src/domains/Common.ts
@@ -1,2 +1,0 @@
-export { default as Header } from "./Header/Header";
-export { default as SubHeader } from "./SubHeader/SubHeader";
